LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-077-expected.txt


FAIL Content-visibility is not animatable assert_equals: expected (string) "visible" but got (undefined) undefined
